+++ etsconfigtool/src/etsconfigtool.cpp
@@ -1,6 +1,6 @@
 #include "etsconfigtool.h"
 #include "../src/knxexception.h"
-#include "../src/knxgroupobject/datapointpool.h"
+#include "../src/knxgroupobject/dptwbmqttbuilder.h"
 #include <algorithm>
 #include <iostream>
 #include <unordered_map>
