+++ src/main.cpp
@@ -1,8 +1,8 @@
 #include "config.h"
 #include "configurator.h"
 #include "knxclientservice.h"
-#include "knxgroupobject/datapointpool.h"
 #include "knxgroupobject/dptjsonbuilder.h"
+#include "knxgroupobject/dptwbmqttbuilder.h"
 #include "knxgroupobject/mqttbuilder.h"
 #include "knxgroupobjectcontroller.h"
 #include "knxlegacydevice.h"
